Move some things around for sake of cleanliness
